// packages/ui/exchange-ui-core/src/utils/exchange/types.ts
export interface IAsset {
    id: string;
    address: string;
    tokenId: string;
    deviceId: string;
    generationFrom: string;
    generationTo: string;
}
// ...
export interface IOrderBookOrderDTO {
    id: string;
    price: number;
    volume: string;
    product: ProductDTO;
    userId: string;
    assetId?: string;
}